lint: move completion helpers out of ctrl
