+++ src/controller/AccountController.ts
@@ -4,7 +4,6 @@ import { Connection, In, Repository } from 'typeorm';
 
 import { CheckRBAC, HandleNoun, HandleVerb } from 'src/controller';
 import { BaseController, ErrorReplyType } from 'src/controller/BaseController';
-import { createCompletion } from 'src/controller/CompletionController';
 import { Controller, ControllerData, ControllerOptions } from 'src/controller/Controller';
 import { Role } from 'src/entity/auth/Role';
 import { Token } from 'src/entity/auth/Token';
@@ -12,6 +11,8 @@ import { User } from 'src/entity/auth/User';
 import { UserRepository } from 'src/entity/auth/UserRepository';
 import { Command, CommandVerb } from 'src/entity/Command';
 import { Clock } from 'src/utils/Clock';
+
+import { createCompletion } from './helpers';
 
 export const NOUN_GRANT = 'grant';
 export const NOUN_ACCOUNT = 'account';
