+++ src/controller/TokenController.ts
@@ -4,11 +4,12 @@ import { Connection, Equal, LessThan, Repository } from 'typeorm';
 
 import { CheckRBAC, HandleNoun, HandleVerb } from 'src/controller';
 import { BaseController } from 'src/controller/BaseController';
-import { createCompletion } from 'src/controller/CompletionController';
 import { Controller, ControllerData, ControllerOptions } from 'src/controller/Controller';
 import { Token } from 'src/entity/auth/Token';
 import { Command, CommandVerb } from 'src/entity/Command';
 import { Clock } from 'src/utils/Clock';
+
+import { createCompletion } from './helpers';
 
 export const NOUN_TOKEN = 'token';
